domain 패키지에 어울리지 않은 요소들 옮기기
