+++ src/main/java/com/myproject/doseoro/adaptor/api/PageController.java
@@ -4,11 +4,11 @@ import com.myproject.doseoro.adaptor.logger.Logging;
 import com.myproject.doseoro.application.book.handler.FindHomeDisplayingBooksCommandHandler;
 import com.myproject.doseoro.application.book.handler.GetLikedBooksByUserQuery;
 import com.myproject.doseoro.application.book.handler.GetAllSaleBooksQuery;
+import com.myproject.doseoro.application.book.vo.HomeDisplayedBookVO;
 import com.myproject.doseoro.application.identity.handler.GetUserInformationQuery;
-import com.myproject.doseoro.domain.book.dto.GetLikedBooksByUserDtoResult;
-import com.myproject.doseoro.domain.book.dto.GetAllSaleBooksResult;
-import com.myproject.doseoro.domain.book.vo.*;
-import com.myproject.doseoro.domain.identity.dto.GetUserInformationDtoResult;
+import com.myproject.doseoro.application.book.dto.GetLikedBooksByUserDtoResult;
+import com.myproject.doseoro.application.book.dto.GetAllSaleBooksResult;
+import com.myproject.doseoro.application.identity.dto.GetUserInformationDtoResult;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
