+++ src/main/java/com/myproject/doseoro/adaptor/infra/mybatis/identity/IdentityMybatisRepository.java
@@ -2,8 +2,8 @@ package com.myproject.doseoro.adaptor.infra.mybatis.identity;
 
 import com.myproject.doseoro.adaptor.infra.dao.DoseoroDao;
 import com.myproject.doseoro.adaptor.global.error.exception.BusinessException;
-import com.myproject.doseoro.domain.identity.abstraction.IdentityRepository;
-import com.myproject.doseoro.domain.identity.vo.*;
+import com.myproject.doseoro.application.identity.vo.*;
+import com.myproject.doseoro.application.abstraction.IdentityRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.password.PasswordEncoder;
