+++ src/main/java/com/myproject/doseoro/application/book/dto/GetLikedBooksByUserDtoResult.java
@@ -1,6 +1,6 @@
-package com.myproject.doseoro.domain.book.dto;
+package com.myproject.doseoro.application.book.dto;
 
-import com.myproject.doseoro.domain.book.vo.BookVO;
+import com.myproject.doseoro.application.book.vo.BookVO;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.ToString;
