+++ src/main/java/com/myproject/doseoro/application/book/handler/GetLikedBooksByUserQuery.java
@@ -2,10 +2,10 @@ package com.myproject.doseoro.application.book.handler;
 
 import com.myproject.doseoro.adaptor.global.util.session.AccessUserSessionManager;
 import com.myproject.doseoro.application.abstraction.CommandQuery;
-import com.myproject.doseoro.domain.book.abstraction.BookRepository;
-import com.myproject.doseoro.domain.book.dto.GetLikedBooksByUserDtoResult;
-import com.myproject.doseoro.domain.book.vo.AllLikedBookVO;
-import com.myproject.doseoro.domain.book.vo.BookVO;
+import com.myproject.doseoro.application.abstraction.BookRepository;
+import com.myproject.doseoro.application.book.dto.GetLikedBooksByUserDtoResult;
+import com.myproject.doseoro.application.book.vo.AllLikedBookVO;
+import com.myproject.doseoro.application.book.vo.BookVO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
