+++ src/main/java/com/myproject/doseoro/application/identity/dto/IdentityDTO.java
@@ -1,12 +1,16 @@
-package com.myproject.doseoro.domain.identity.vo;
+package com.myproject.doseoro.application.identity.dto;
 
+import lombok.AccessLevel;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import javax.validation.constraints.NotBlank;
+import java.io.Serializable;
 
 @Getter
-public class IdentityVO {
+public class IdentityDTO implements Serializable {
+
+    @NotBlank
     private final String id;
     private final String email;
     private final String password;
@@ -16,9 +20,7 @@ public class IdentityVO {
     private final String forgotPwQuestion;
     private final String forgotPwAnswer;
 
-    private final HttpServletResponse response;
-
-    public IdentityVO(String id, String email, String password, String name, String nickName, String phone, String forgotPwQuestion, String forgotPwAnswer, HttpServletResponse response) {
+    public IdentityDTO(String id, String email, String password, String name, String nickName, String phone, String forgotPwQuestion, String forgotPwAnswer) {
         this.id = id;
         this.email = email;
         this.password = password;
@@ -27,12 +29,11 @@ public class IdentityVO {
         this.phone = phone;
         this.forgotPwQuestion = forgotPwQuestion;
         this.forgotPwAnswer = forgotPwAnswer;
-        this.response = response;
     }
 
     @Override
     public String toString() {
-        return "IdentityVO{" +
+        return "IdentityDto{" +
                 "id='" + id + '\'' +
                 ", email='" + email + '\'' +
                 ", password='" + password + '\'' +
@@ -41,14 +42,6 @@ public class IdentityVO {
                 ", phone='" + phone + '\'' +
                 ", forgotPwQuestion='" + forgotPwQuestion + '\'' +
                 ", forgotPwAnswer='" + forgotPwAnswer + '\'' +
-                ", response=" + response +
                 '}';
     }
-
-    //    public IdentityVO viewUser(IdentityVO vo, IdentityMybatisService repository) {
-//        System.out.println("view User = "+vo);
-//        IdentityVO a = repository.findByEmail(vo.getEmail());
-//        System.out.println("@@@@@@@@ = "+a);
-//        return a;
-//    }
 }
