+++ src/main/java/com/myproject/doseoro/application/identity/handler/AuthenticateUserCommandHandler.java
@@ -2,10 +2,10 @@ package com.myproject.doseoro.application.identity.handler;
 
 import com.myproject.doseoro.adaptor.logger.Logging;
 import com.myproject.doseoro.application.abstraction.CommandHandler;
-import com.myproject.doseoro.domain.identity.abstraction.IdentityRepository;
-import com.myproject.doseoro.domain.identity.vo.AccessUserVO;
+import com.myproject.doseoro.application.abstraction.IdentityRepository;
+import com.myproject.doseoro.application.identity.vo.AccessUserVO;
 import com.myproject.doseoro.adaptor.global.util.session.AccessUserSessionManager;
-import com.myproject.doseoro.domain.identity.vo.IdentityVO;
+import com.myproject.doseoro.application.identity.vo.IdentityVO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
