+++ src/main/java/com/myproject/doseoro/application/identity/vo/IdentityVO.java
@@ -1,16 +1,12 @@
-package com.myproject.doseoro.domain.identity.dto;
+package com.myproject.doseoro.application.identity.vo;
 
-import lombok.AccessLevel;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
-import javax.validation.constraints.NotBlank;
-import java.io.Serializable;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 
 @Getter
-public class IdentityDTO implements Serializable {
-
-    @NotBlank
+public class IdentityVO {
     private final String id;
     private final String email;
     private final String password;
@@ -20,7 +16,9 @@ public class IdentityDTO implements Serializable {
     private final String forgotPwQuestion;
     private final String forgotPwAnswer;
 
-    public IdentityDTO(String id, String email, String password, String name, String nickName, String phone, String forgotPwQuestion, String forgotPwAnswer) {
+    private final HttpServletResponse response;
+
+    public IdentityVO(String id, String email, String password, String name, String nickName, String phone, String forgotPwQuestion, String forgotPwAnswer, HttpServletResponse response) {
         this.id = id;
         this.email = email;
         this.password = password;
@@ -29,11 +27,12 @@ public class IdentityDTO implements Serializable {
         this.phone = phone;
         this.forgotPwQuestion = forgotPwQuestion;
         this.forgotPwAnswer = forgotPwAnswer;
+        this.response = response;
     }
 
     @Override
     public String toString() {
-        return "IdentityDto{" +
+        return "IdentityVO{" +
                 "id='" + id + '\'' +
                 ", email='" + email + '\'' +
                 ", password='" + password + '\'' +
@@ -42,6 +41,14 @@ public class IdentityDTO implements Serializable {
                 ", phone='" + phone + '\'' +
                 ", forgotPwQuestion='" + forgotPwQuestion + '\'' +
                 ", forgotPwAnswer='" + forgotPwAnswer + '\'' +
+                ", response=" + response +
                 '}';
     }
+
+    //    public IdentityVO viewUser(IdentityVO vo, IdentityMybatisService repository) {
+//        System.out.println("view User = "+vo);
+//        IdentityVO a = repository.findByEmail(vo.getEmail());
+//        System.out.println("@@@@@@@@ = "+a);
+//        return a;
+//    }
 }
