+++ src/test/java/com/myproject/doseoro/packages/book/handler/RegisterBookCommandHandlerTest.java
@@ -5,8 +5,8 @@ import com.myproject.doseoro.adaptor.infra.mybatis.book.BookMybatisRepository;
 import com.myproject.doseoro.adaptor.infra.mybatis.identity.IdentityMybatisRepository;
 import com.myproject.doseoro.application.book.handler.RegisterBookCommandHandler;
 import com.myproject.doseoro.application.identity.handler.CreateUserIdentityCommandHandler;
-import com.myproject.doseoro.domain.book.vo.RegisterBookVO;
-import com.myproject.doseoro.domain.identity.vo.SignUpVO;
+import com.myproject.doseoro.application.book.vo.RegisterBookVO;
+import com.myproject.doseoro.application.identity.vo.SignUpVO;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
