+++ src/test/java/com/myproject/doseoro/packages/identity/handler/AuthenticateUserCommandHandlerTest.java
@@ -6,8 +6,8 @@ import com.myproject.doseoro.adaptor.global.error.exception.BusinessException;
 import com.myproject.doseoro.adaptor.global.util.session.AccessUserSessionManager;
 import com.myproject.doseoro.application.identity.handler.AuthenticateUserCommandHandler;
 import com.myproject.doseoro.application.identity.handler.CreateUserIdentityCommandHandler;
-import com.myproject.doseoro.domain.identity.vo.IdentityVO;
-import com.myproject.doseoro.domain.identity.vo.SignUpVO;
+import com.myproject.doseoro.application.identity.vo.IdentityVO;
+import com.myproject.doseoro.application.identity.vo.SignUpVO;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
