+++ src/test/java/com/myproject/doseoro/packages/identity/handler/CreateUserIdentityCommandHandlerTest.java
@@ -4,7 +4,7 @@ import com.myproject.doseoro.adaptor.infra.dao.DoseoroDao;
 import com.myproject.doseoro.adaptor.infra.mybatis.identity.IdentityMybatisRepository;
 import com.myproject.doseoro.adaptor.global.error.exception.BusinessException;
 import com.myproject.doseoro.application.identity.handler.CreateUserIdentityCommandHandler;
-import com.myproject.doseoro.domain.identity.vo.SignUpVO;
+import com.myproject.doseoro.application.identity.vo.SignUpVO;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
